move to own folder & clean up
